Tiny change to scaleFactor; normally I wouldn't commit this, but it's night and I'll need it tomorrow to avoid merge conflict

+++ StarshipArena/src/Starship.java
@@ -40,7 +40,8 @@ public class Starship {
 	Point[] points;
 	Point[] hitbox;
 	
-	static int scaleFactor = 2;
+	//Change scale of ships
+	static double scaleFactor = 2;
 	
 	String team;
 	int control_group = 0;
